zDYNAMIC_CAST fix
Added new template class U to zDYNAMIC_CAST to check if T and pObject have unique classdefs.
Removed C style casts to avoid force casting.
Added static_assert to make sure if T and U is a base of zCObject.
With C++17 the check is guaranteed to be compile time evaluated.