Fixed issue with zero initializing oCNpc::oSDamageDescriptor:
The issue is that by default every member of `oCNpc::oSDamageDescriptor` ise set to 0, or `nullptr`. It's important to mimic the original behaviour from the game code, and we can't do that without using `memcpy` or `memset` (which are hacks), because Union explicitly defines default constructor that does nothing (in fact it shadows the zero initializer constructor).
To prevent that, i've added `= default;` statement, so that zero initializer constructor can be used by doing `= {}`, which solves problems where oCNpc::oSDamageDescriptor is declared on the stack.